% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBF-package.R
\docType{package}
\name{DBF-package}
\alias{DBF}
\alias{DBF-package}
\title{DBF: Provide Dummy DB Syntax in Namespace to Use Functions with dplyr}
\description{
Functions of the dplyr package has the ability 
            to convert database-specific functions directly into sql 
            when the database is the source. 
            However, you can only use the web to search its features. 
            If you register function names in RDB as dummy functions in R, 
            it will be easier to use. DBF is an abbreviation for DataBase Function, 
            which provides convenience when using database-specific functions in tidyverse syntax.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/DBF}
  \item Report bugs at \url{https://github.com/mrchypark/DBF/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
