% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_function.R
\name{set_function}
\alias{set_function}
\title{set db function to dummy function in R}
\usage{
set_function(list_of_function, ns_name)
}
\arguments{
\item{ns_name}{namespace name to set. Defualt is  like postgres.}

\item{list}{list of function generated using \link{get_function}.}
}
\description{
set db function name to dummy object in R. now assign NULL
}
